Move test util to test dir
